Refactoring, moving json to a separate libX.[ch] unit
